requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0

# UI & Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0
